% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prufer.R
\name{all_prufer_sequences}
\alias{all_prufer_sequences}
\title{Generate all Pr端fer sequences}
\usage{
all_prufer_sequences(n)
}
\arguments{
\item{n}{Number of variables, n >= 3 (n = 2 is the empty Pr端fer sequence)}
}
\value{
A matrix with n^(n-2) rows and each row is a Pr端fer sequence
}
\description{
Generate all Pr端fer sequences
}
\examples{
all_prufer_sequences(3)

}
