% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry-all-directly.R
\name{all_tcherries}
\alias{all_tcherries}
\title{Generate all k'th order t-cherry junction trees}
\usage{
all_tcherries(n, k, remove_duplicates = TRUE, verbose = FALSE)
}
\arguments{
\item{n}{Number of variables}

\item{k}{Order of t-cherry junction tree (clique size)}
}
\description{
Generate all k'th order t-cherry junction trees
}
\examples{
all_tcherries(4, 2)

}
