% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry-generate-all.R
\name{all_tcherries_cpp_pure}
\alias{all_tcherries_cpp_pure}
\title{Now, extend by this strategy:}
\usage{
all_tcherries_cpp_pure(n, k, remove_duplicates = TRUE, verbose = FALSE)
}
\description{
With RIP (running intersection property) in mind,
Take last clique of size k and take all (k-1) subsets
together with each of of the unused variables, hence
there are choose(k, k-1)*length(unused) new possibilities
}
