% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry-all-directly.R
\name{all_tcherries_r}
\alias{all_tcherries_r}
\title{Now, extend by this strategy:}
\usage{
all_tcherries_r(n, k, verbose = FALSE)
}
\description{
With RIP (running intersection property) in mind,
Take last clique of size k and take all (k-1) subsets
together with each of of the unused variables, hence
there are choose(k, k-1)*length(unused) new possibilities
}
