% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndgm.R
\docType{package}
\name{learndgm-package}
\alias{learndgm}
\alias{learndgm-package}
\title{Learn structures of decomposable graphical models from data}
\description{
Learn structures of decomposable graphical models from data
}
\details{
See vignettes and manual for documentation.

Disclaimer:
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, TITLE AND NON-INFRINGEMENT.
IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR ANYONE DISTRIBUTING
THE SOFTWARE BE LIABLE FOR ANY DAMAGES OR OTHER LIABILITY, WHETHER IN CONTRACT,
TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mikldk/learndgm/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Meyer Andersen \email{mikl@math.aau.dk}

}
