% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prufer.R
\name{prufer_sequences_as_graphs}
\alias{prufer_sequences_as_graphs}
\title{Convert Prüfer sequences in matrix rows to graphs}
\usage{
prufer_sequences_as_graphs(s)
}
\value{
List with graphs
}
\description{
Convert Prüfer sequences in matrix rows to graphs
}
