% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry-all-directly.R
\name{remove_equal_models}
\alias{remove_equal_models}
\title{Remove entries that give the same model}
\usage{
remove_equal_models(models)
}
\arguments{
\item{models}{List of models, e.g. from \code{\link[=all_tcherries]{all_tcherries()}}}
}
\description{
Currently by constructing adjacency matrix and
checking equality.
}
\details{
Only models that are finished.
}
\examples{
m <- all_tcherries(4, 2)
remove_equal_models(m)

}
